# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: id,asc
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
      operationId: getOrders
      description: Get orders as a paginated collection
    post:
      summary: Create Order
      tags:
        - store
      operationId: createOrder
      description: Create a new order with an existing customer and one or more existing products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Customer or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /order/{id}:
    get:
      summary: Get Order By ID
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      operationId: getOrderById
      description: Get a specific order by ID
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Optional case-insensitive substring search within a single word of the customer name
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerSummary'
        '400':
          description: Invalid customer request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      operationId: getCustomers
      description: Get all customers, or filter them by a query string against name words
    post:
      summary: Create Customer
      tags:
        - store
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid customer request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      operationId: createCustomer
      description: Create a new customer
  /products:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Invalid product request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      operationId: getProducts
      description: Get all products
    post:
      summary: Create Product
      tags:
        - store
      operationId: createProduct
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid product request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /products/{id}:
    get:
      summary: Get Product By ID
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      operationId: getProductById
      description: Get a specific product by ID
components:
  schemas:
    ApiError:
      title: ApiError
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    CreateOrderRequest:
      title: CreateOrderRequest
      type: object
      required:
        - description
        - customerId
        - productIds
      properties:
        description:
          type: string
        customerId:
          type: integer
          format: int64
        productIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
    CreateProductRequest:
      title: CreateProductRequest
      type: object
      required:
        - description
      properties:
        description:
          type: string
    OrderProduct:
      title: OrderProduct
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProduct'
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
    CustomerSummary:
      title: CustomerSummary
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    OrderSummary:
      title: OrderSummary
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProduct'
    OrderPage:
      title: OrderPage
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummary'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
tags:
  - name: store
    description: store operations